refactors
- simplify docs js/css imports into docs.js and docs.css
- add details-utils wrapper for persistent details/summary elements
- add api/ for multiple API versions

+++ bin/build.sh
@@ -3,7 +3,11 @@
 set -e
 
 mkdir -p dist
-rm -r dist/* || true
+
+if [ $NODE_ENV != "development" ]
+then
+  rm -r dist/* || true
+fi
 
 PATH_PREFIX=${PATH_PREFIX:-/}
